Allow multiple dynamic classes
This update allows for multiple dynamic classes. For example, if you have the following class names `full-screen color-scheme-light`  and change it to `full-screen color-scheme-dark`, currently it will continue to append each new class string onto the body classes because the entire string is not a match.  (e.g. `full-screen color-scheme-light full-screen color-scheme-dark`)

Placing the names to remove first ensures that we don't delete a duplicate class name.