Replace `set-body-class` component with helper
Helpers automatically get called when their inputs change, which makes this implementation quite a bit easier and removes the need for an observer.

When the helper is first called it generates a unique id, and registers itself with the id and the current value on the service. When the value changes it registers again with the same id to update the existing value. Finally, when the helper goes out of scope and is destroyed it unregisters from the service again. These service changes trigger the `updateBodyClass()` on the next runloop, which updates the `class` attribute based on all the registrations.